멀티플레이어 관련
[/] 멀티플레이어 업데이트 함수 호출 안하던 점 수정
[/] 애니메이션 상태를 네임스페이스를 사용한 형태로 변경

+++ 소스/Client/texture.cpp
@@ -38,7 +38,7 @@ void Texture::CreateTexture(const ComPtr<ID3D12Device>& device)
 	// 셰이더 리소스 뷰 서술자 생성
 	if (m_srvHeap) m_srvHeap.Reset();
 	D3D12_DESCRIPTOR_HEAP_DESC srvHeapDesc{};
-	srvHeapDesc.NumDescriptors = m_textures.size(); // SRV 개수
+	srvHeapDesc.NumDescriptors = static_cast<UINT>(m_textures.size()); // SRV 개수
 	srvHeapDesc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
 	srvHeapDesc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;
 	device->CreateDescriptorHeap(&srvHeapDesc, IID_PPV_ARGS(&m_srvHeap));
